rename wrapper integration test package
